fix: move tests to tests/ and replace double-click sound with single click
- Move DopplerCalculator.test.ts and WaveGenerator.test.ts from src/ to
  tests/ — the compliance check requires tests/ not src/ (was: FAIL)
- Update relative imports in test files to ../src/doppler-effect/model/
- Update vitest.config.ts include pattern to tests/**/*.test.ts
- Add biome.json override for tests/ with relaxed noConsole rule
- Replace assets/click.wav: was a stereo 24-bit WAVE_FORMAT_EXTENSIBLE
  file containing two audible clicks; now a clean mono 16-bit 44100 Hz
  single click (1.5 kHz sine burst, 1 ms attack, 8 ms decay, 60 ms total)
